+++ packages/file-storage/src/fs.ts
@@ -1,0 +1,1 @@
+export { createFsFileStorage } from './lib/backends/fs.ts'