+++ packages/file-storage/src/local.ts
@@ -1,1 +1,0 @@
-export { createLocalFileStorage } from './lib/backends/local.ts'